+++ packages/s2-core/src/sheet-type/spread-sheet.ts
@@ -22,7 +22,6 @@ import {
 import { BaseCell } from '../cell';
 import {
   InterceptType,
-  LayoutWidthType,
   S2Event,
   getDefaultSeriesNumberText,
   getTooltipOperatorSortMenus,
@@ -608,14 +607,14 @@ export abstract class SpreadSheet extends EE {
   }
 
   public override on<K extends keyof EmitterType>(
-    event: K,
+    event: `${K}`,
     listener: EmitterType[K],
   ): this {
     return super.on(event, listener);
   }
 
   public override emit<K extends keyof EmitterType>(
-    event: K,
+    event: `${K}`,
     ...args: Parameters<EmitterType[K]>
   ): void {
     super.emit(event, ...args);
@@ -652,7 +651,7 @@ export abstract class SpreadSheet extends EE {
     };
   }
 
-  public getLayoutWidthType(): LayoutWidthType {
+  public getLayoutWidthType() {
     return this.options.style?.layoutWidthType!;
   }