+++ packages/s2-react-components/__tests__/unit/components/config/frozen-panel/frozen-input-number/index-spec.tsx
@@ -1,0 +1,24 @@
+import { FrozenInputNumber } from '@/components';
+import { render } from '@testing-library/react';
+import React from 'react';
+
+describe('Frozen Input Number Component Tests', () => {
+  test('should render correctly panel', () => {
+    const { asFragment } = render(<FrozenInputNumber />);
+
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  test('should set custom input number props', () => {
+    const { asFragment } = render(
+      <FrozenInputNumber
+        size="large"
+        step={2}
+        className="test"
+        style={{ color: 'red' }}
+      />,
+    );
+
+    expect(asFragment()).toMatchSnapshot();
+  });
+});